requirements.txt
beautifulsoup4==4.9.1
Pillow==7.2.0
requests-html==0.10.0
urllib3==1.25.9
tqdm==4.45.0
